requirement.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0

certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
mongoengine==0.26.0
multidict==6.0.4
openai==0.26.1
Pillow==9.4.0
PyJWT==2.6.0
pymongo==3.12.3
pytz==2022.7.1
requests==2.28.2
sqlparse==0.2.4
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==3.2.18
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
importlib-metadata==6.0.0
install==1.3.5
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==3.0.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
zipp==3.13.0
